feat (RenderingEngine) add the menu of the RenderingEngine before lauching the game
